+++ example-apps/src/features/kyc-airdrop/utils/blockchain.schema.ts
@@ -1,5 +1,5 @@
 import type { GetTxAuthDataSignatureResponse } from "@nexeraprotocol/identity-schemas/identity";
 
-export type MintResponse = {
+export type ClaimResponse = {
   signatureResponse: GetTxAuthDataSignatureResponse;
 };
